requirements.txt
certifi==2018.8.24
chardet==3.0.4
dataclasses==0.6
Django==2.1.2
djangorestframework==3.8.2
djongo==1.2.30
idna==2.7
oauthlib==2.1.0
pymongo==3.7.1
PySocks==1.6.8
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.2.4
tweepy==3.6.0
urllib3==1.23
